+++ src/pixel.rs
@@ -1,4 +1,4 @@
-/// A RGBA or BGRA pixel.
+/// A 32-bit pixel with 4 components.
 ///
 /// # Representation
 ///
@@ -8,6 +8,13 @@
 /// type have the same in-memory representation as a `u32`.
 ///
 /// [`PixelFormat::default()`]: crate::PixelFormat#default
+///
+/// # Default
+///
+/// The [`Default`] impl returns a transparent black pixel. Beware that this might not be what you
+/// want if using [`AlphaMode::Opaque`] (which is the default).
+///
+/// [`AlphaMode::Opaque`]: crate::AlphaMode::Opaque
 ///
 /// # Example
 ///
@@ -20,10 +27,10 @@
 /// assert_eq!(red.r, 255);
 /// assert_eq!(red.g, 128);
 /// assert_eq!(red.b, 0);
-/// # // assert_eq!(red.a, 0xff);
-/// #
-/// # // let from_struct_literal = Pixel { r: 255, g: 0x80, b: 0, a: 0xff };
-/// # // assert_eq!(red, from_struct_literal);
+/// assert_eq!(red.a, 0xff);
+///
+/// let from_struct_literal = Pixel { r: 255, g: 0x80, b: 0, a: 0xff };
+/// assert_eq!(red, from_struct_literal);
 /// ```
 ///
 /// Convert a pixel to an array of `u8`s.
@@ -36,8 +43,8 @@
 /// let red = unsafe { core::mem::transmute::<Pixel, [u8; 4]>(red) };
 ///
 /// match PixelFormat::default() {
-///     PixelFormat::Bgrx => assert_eq!(red[2], 255),
-///     PixelFormat::Rgbx => assert_eq!(red[0], 255),
+///     PixelFormat::Bgra => assert_eq!(red[2], 255),
+///     PixelFormat::Rgba => assert_eq!(red[0], 255),
 /// }
 /// ```
 ///
@@ -51,8 +58,8 @@
 /// let red = unsafe { core::mem::transmute::<Pixel, u32>(red) };
 ///
 /// match PixelFormat::default() {
-///     PixelFormat::Bgrx => assert_eq!(red, u32::from_ne_bytes([0x00, 0x00, 0xff, 0x00])),
-///     PixelFormat::Rgbx => assert_eq!(red, u32::from_ne_bytes([0xff, 0x00, 0x00, 0x00])),
+///     PixelFormat::Bgra => assert_eq!(red, u32::from_ne_bytes([0x00, 0x00, 0xff, 0xff])),
+///     PixelFormat::Rgba => assert_eq!(red, u32::from_ne_bytes([0xff, 0x00, 0x00, 0xff])),
 /// }
 /// ```
 #[repr(C)]
@@ -82,14 +89,12 @@ pub struct Pixel {
     ///
     /// `0xff` here means opaque, whereas `0` means transparent.
     ///
-    /// NOTE: Transparency is not yet supported, see [#17], so this doesn't actually do anything.
-    ///
-    /// [#17]: https://github.com/rust-windowing/softbuffer/issues/17
-    pub(crate) a: u8,
+    /// Make sure to set this correctly according to the [`AlphaMode`][crate::AlphaMode].
+    pub a: u8,
 }
 
 impl Pixel {
-    /// Create a new pixel from a red, a green and a blue component.
+    /// Create a new opaque pixel from a red, a green and a blue component.
     ///
     /// # Example
     ///
@@ -100,11 +105,10 @@ impl Pixel {
     /// assert_eq!(red.r, 255);
     /// ```
     pub const fn new_rgb(r: u8, g: u8, b: u8) -> Self {
-        // FIXME(madsmtm): Change alpha to `0xff` once we support transparency.
-        Self { r, g, b, a: 0x00 }
+        Self { r, g, b, a: 0xff }
     }
 
-    /// Create a new pixel from a blue, a green and a red component.
+    /// Create a new opaque pixel from a blue, a green and a red component.
     ///
     /// # Example
     ///
@@ -115,11 +119,38 @@ impl Pixel {
     /// assert_eq!(red.r, 255);
     /// ```
     pub const fn new_bgr(b: u8, g: u8, r: u8) -> Self {
-        // FIXME(madsmtm): Change alpha to `0xff` once we support transparency.
-        Self { r, g, b, a: 0x00 }
+        Self { r, g, b, a: 0xff }
     }
 
-    // TODO(madsmtm): Once we have transparency, add `new_rgba` and `new_bgra` methods.
+    /// Create a new pixel from a red, a green, a blue and an alpha component.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// # use softbuffer::Pixel;
+    /// #
+    /// let red = Pixel::new_rgba(0xff, 0, 0, 0x7f);
+    /// assert_eq!(red.r, 255);
+    /// assert_eq!(red.a, 127);
+    /// ```
+    pub const fn new_rgba(r: u8, g: u8, b: u8, a: u8) -> Self {
+        Self { r, g, b, a }
+    }
+
+    /// Create a new pixel from a blue, a green, a red and an alpha component.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// # use softbuffer::Pixel;
+    /// #
+    /// let red = Pixel::new_bgra(0, 0, 0xff, 0x7f);
+    /// assert_eq!(red.r, 255);
+    /// assert_eq!(red.a, 127);
+    /// ```
+    pub const fn new_bgra(b: u8, g: u8, r: u8, a: u8) -> Self {
+        Self { r, g, b, a }
+    }
 }
 
 // TODO: Implement `Add`/`Mul`/similar `std::ops` like `rgb` does?